chore: add logging
