+++ packages/react-app-revamp/config/wagmi/index.ts
@@ -21,8 +21,7 @@ import {
   walletConnectWallet,
   zerionWallet,
 } from "@rainbow-me/rainbowkit/wallets";
-import { Transport } from "viem";
-import { cookieStorage, createConfig, createStorage, fallback, http } from "wagmi";
+import { cookieStorage, createConfig, createStorage } from "wagmi";
 import { arbitrumOne } from "./custom-chains/arbitrumOne";
 import { avalanche } from "./custom-chains/avalanche";
 import { base } from "./custom-chains/base";
@@ -51,16 +50,13 @@ import { swell } from "./custom-chains/swell";
 import { unichain } from "./custom-chains/unichain";
 import { zora } from "./custom-chains/zora";
 import { isParaWalletConfigured, paraWallet } from "./para";
+import createTransports from "./transports";
 
 declare module "wagmi";
 
 type ChainImages = {
   [key: string]: string;
 };
-
-const isProduction = process.env.NODE_ENV === "production";
-
-type Transports = Record<Chain["id"], Transport>;
 
 export const chains: readonly [Chain, ...Chain[]] = [
   polygon,
@@ -139,30 +135,6 @@ const connectors = connectorsForWallets(
     appName: appName,
   },
 );
-
-const createTransports = (chains: readonly [Chain, ...Chain[]]): Transports => {
-  const headers = isProduction
-    ? { Referer: "https://jokerace.io/" }
-    : { Referer: "https://jokerace-git-chore-fix-referer-implementation-jokerace.vercel.app/" };
-
-  return chains.reduce<Transports>((acc, chain) => {
-    if (chain.rpcUrls?.default?.http?.[0] && chain.rpcUrls?.public?.http?.[0]) {
-      acc[chain.id] = fallback([
-        http(chain.rpcUrls.default.http[0], {
-          fetchOptions: {
-            headers,
-          },
-        }),
-        http(chain.rpcUrls.public.http[0], {
-          fetchOptions: {
-            headers,
-          },
-        }),
-      ]);
-    }
-    return acc;
-  }, {});
-};
 
 const transports = createTransports(chains);
